% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{box_plot}
\alias{box_plot}
\title{Boxplot}
\usage{
box_plot(data, xvar, yvar, xlab = "", ylab = "", facet = FALSE)
}
\arguments{
\item{data}{Cleaned data}

\item{xvar}{Variable on x-axis}

\item{yvar}{Variable on y-axis}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{facet}{Split the plot by visit number}
}
\value{
Plot
}
\description{
Boxplot
}
\examples{
ds \%>\%
  filter(fVN == "Baseline") \%>\%
  select(acrStatus, udbpCrRatio) \%>\%
  na.omit() \%>\%
  box_plot("acrStatus", "log(udbpCrRatio)",
           "Albuminuria",
           "log uVDBP:Creatinine")
}
