% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{box_plot_poster}
\alias{box_plot_poster}
\title{Boxplots where the colour theme matches ada poster}
\usage{
box_plot_poster(data, xvar, yvar, xlab = "", ylab = "", facet = FALSE)
}
\arguments{
\item{facet}{}
}
\value{
Plot
}
\description{
Boxplots where the colour theme matches ada poster
}
