% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{box_plot_slides}
\alias{box_plot_slides}
\title{Boxplots where the colour theme matches health template slides}
\usage{
box_plot_slides(data, xvar, yvar, xlab = "", ylab = "", facet = FALSE)
}
\arguments{
\item{facet}{}
}
\value{
Plot
}
\description{
Boxplots where the colour theme matches health template slides
}
