% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\title{Fetch data from the original source}
\usage{
fetch_data()
}
\value{
Saves the wrangled data into the data/ folder.
}
\description{
This function fetchs the main dataset, keeps variables relevant to
the analysis, restrict the sample size as needed, and lastly save
the new dataset as an `.RData` file.
}
\examples{
fetch_data()

}
