% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{histo_plot}
\alias{histo_plot}
\title{Histogram}
\usage{
histo_plot(data, variable, bin, xlab = "", facet = FALSE)
}
\arguments{
\item{data}{Cleaned data}

\item{variable}{Variable of interest}

\item{bin}{Width of bars}

\item{xlab}{x-axis label}

\item{facet}{Separate the graph by visit number}
}
\value{
Plot
}
\description{
Histogram
}
