% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{line_plot}
\alias{line_plot}
\title{Line Plot}
\usage{
line_plot(data, xvar, yvar, byvar, xlab = "", ylab = "")
}
\arguments{
\item{data}{Dataset (cleaned)}

\item{xvar}{Variable on x axis (continous or factor)}

\item{yvar}{Variable on y axis (continuous or factor)}

\item{xlab}{Axis label for x axis}

\item{ylab}{Axis label for y axis}
}
\value{
Plot
}
\description{
Line Plot
}
\examples{
ds \%>\%
  dplyr::filter(SID < 1050 ) \%>\%
  line_plot("fVN", "UDBP", "SID", "Visit Number", "UDBP")
}
