% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meds.R
\name{med_ace}
\alias{med_ace}
\title{Detailed classification of medication data to include ACE inhibitors}
\usage{
med_ace()
}
\description{
Detailed classification of medication data to include ACE inhibitors
}
\examples{

}
