% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{prep_gee_data}
\alias{prep_gee_data}
\title{Prepare the project data for analysis through GEE}
\usage{
prep_gee_data(data)
}
\arguments{
\item{data}{project data}
}
\description{
Prepare the project data for analysis through GEE
}
