% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename_table_rows}
\alias{rename_table_rows}
\title{Renames variables in data set to improve understanding and include units}
\usage{
rename_table_rows(x)
}
\description{
Renames variables in data set to improve understanding and include units
}
