% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_plot_poster}
\alias{scatter_plot_poster}
\title{Scatterplot where the colour theme matches ada poster}
\usage{
scatter_plot_poster(data, xvar, yvar, xlab = "", ylab = "", line = TRUE,
  facet = FALSE)
}
\arguments{
\item{data}{Clearned data}

\item{xvar}{Variable on x-axis}

\item{yvar}{Variable on y-axis}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{line}{Line of best fit}

\item{facet}{Split the plot by visit number}
}
\value{
Plot
}
\description{
Scatterplot where the colour theme matches ada poster
}
